refactored and updated
