+++ scheduler.py
@@ -3,12 +3,12 @@
 import os
 import sys
 
-from tasks.1_retrieve_forecasts_from_BOM import retrieve_forecasts
-from tasks.2_check_db_integrity import integrity_check
-from tasks.3_forecast_dataframe import build_forecast_dataframe
-from tasks.4_generate_heatmaps import generate_heatmaps
-from training_data.1_training_weather import build_training_dataframe
-from tasks.5_prophet_forecaster import build_prophet_forecaster
+from tasks.task1_retrieve_forecasts_from_BOM import retrieve_forecasts
+from tasks.task2_check_db_integrity import integrity_check
+from tasks.task3_forecast_dataframe import build_forecast_dataframe
+from tasks.task4_generate_heatmaps import generate_heatmaps
+from training_data.task1_training_weather import build_training_dataframe
+from tasks.task5_prophet_forecaster import build_prophet_forecaster
 
 # Runs the following tasks daily.
 def main():
@@ -28,13 +28,13 @@ def main():
     generate_heatmaps()  # Train, fit and save prophet forecasts.
     print('LOG: Complete - Predict Prophet Forecasts', '\n')
 
-    print('LOG: Starting - Build Training Dataframe')
-    build_training_dataframe()  # Build and save the training data.
-    print('LOG: Complete - Build Training Dataframe')
+#     print('LOG: Starting - Build Training Dataframe')
+#     build_training_dataframe()  # Build and save the training data.
+#     print('LOG: Complete - Build Training Dataframe')
 
-    print('LOG: Starting - Predict Prophet Forecasts')
-    build_prophet_forecaster()  # Train, fit and save prophet forecasts.
-    print('LOG: Complete - Predict Prophet Forecasts', '\n')
+#     print('LOG: Starting - Predict Prophet Forecasts')
+#     build_prophet_forecaster()  # Train, fit and save prophet forecasts.
+#     print('LOG: Complete - Predict Prophet Forecasts', '\n')
     
     print('LOG: Scheduler Complete')
