+++ static/data/accuracy_dataframe.csv
@@ -221,9 +221,76 @@
 2021-05-19,0.0,0.0,0.0,0.0,1.0,0.0,0.0
 2021-05-20,0.0,0.0,0.0,0.0,-1.0,1.0,-1.0
 2021-05-21,0.0,0.0,0.0,0.0,0.0,-1.0,0.0
-2021-05-22,0.0,0.0,0.0,1.0,0.0,0.0,
-2021-05-23,0.0,0.0,0.0,0.0,0.0,,
-2021-05-24,0.0,0.0,0.0,0.0,,,
-2021-05-25,0.0,0.0,0.0,,,,
-2021-05-26,0.0,0.0,,,,,
-2021-05-27,0.0,,,,,,
+2021-05-22,0.0,0.0,0.0,1.0,0.0,0.0,0.0
+2021-05-23,0.0,0.0,0.0,0.0,0.0,0.0,0.0
+2021-05-24,0.0,0.0,0.0,0.0,-1.0,-1.0,-1.0
+2021-05-25,0.0,0.0,0.0,0.0,0.0,-1.0,0.0
+2021-05-26,0.0,0.0,-1.0,1.0,0.0,0.0,0.0
+2021-05-27,0.0,0.0,0.0,0.0,1.0,0.0,2.0
+2021-05-28,0.0,0.0,-1.0,1.0,0.0,1.0,1.0
+2021-05-29,0.0,0.0,0.0,0.0,1.0,0.0,-2.0
+2021-05-30,0.0,0.0,0.0,1.0,1.0,-1.0,-1.0
+2021-05-31,0.0,-1.0,0.0,0.0,-1.0,-2.0,1.0
+2021-06-01,0.0,0.0,-1.0,0.0,-1.0,1.0,2.0
+2021-06-02,0.0,-1.0,-1.0,-2.0,1.0,1.0,-1.0
+2021-06-03,0.0,0.0,-1.0,1.0,1.0,0.0,1.0
+2021-06-04,0.0,-1.0,1.0,0.0,-2.0,1.0,3.0
+2021-06-05,0.0,0.0,0.0,0.0,1.0,2.0,1.0
+2021-06-06,0.0,0.0,0.0,1.0,0.0,0.0,0.0
+2021-06-07,0.0,0.0,1.0,1.0,0.0,0.0,-1.0
+2021-06-08,0.0,0.0,1.0,0.0,1.0,0.0,-1.0
+2021-06-09,0.0,0.0,1.0,0.0,-1.0,0.0,0.0
+2021-06-10,0.0,1.0,1.0,-1.0,0.0,1.0,-1.0
+2021-06-11,0.0,0.0,-1.0,0.0,0.0,-1.0,1.0
+2021-06-12,0.0,-1.0,-1.0,0.0,-2.0,0.0,0.0
+2021-06-13,0.0,-1.0,0.0,-1.0,0.0,0.0,1.0
+2021-06-14,0.0,0.0,0.0,0.0,1.0,1.0,1.0
+2021-06-15,0.0,-1.0,0.0,0.0,1.0,1.0,0.0
+2021-06-16,0.0,0.0,-1.0,0.0,1.0,0.0,1.0
+2021-06-17,0.0,0.0,0.0,1.0,0.0,1.0,0.0
+2021-06-18,0.0,0.0,1.0,0.0,1.0,0.0,1.0
+2021-06-19,0.0,0.0,1.0,1.0,0.0,0.0,-3.0
+2021-06-20,0.0,0.0,0.0,-1.0,1.0,-3.0,-1.0
+2021-06-21,0.0,0.0,0.0,0.0,-1.0,-1.0,1.0
+2021-06-22,0.0,0.0,0.0,-1.0,-1.0,1.0,1.0
+2021-06-23,0.0,0.0,-1.0,-1.0,1.0,0.0,-1.0
+2021-06-24,0.0,-1.0,0.0,1.0,0.0,0.0,-1.0
+2021-06-25,0.0,0.0,0.0,0.0,0.0,-1.0,1.0
+2021-06-26,0.0,0.0,0.0,0.0,-1.0,0.0,1.0
+2021-06-27,0.0,0.0,0.0,-1.0,1.0,3.0,-3.0
+2021-06-28,0.0,0.0,-1.0,1.0,3.0,-1.0,0.0
+2021-06-29,0.0,0.0,1.0,3.0,-2.0,0.0,0.0
+2021-06-30,0.0,1.0,0.0,-2.0,1.0,0.0,-1.0
+2021-07-01,0.0,1.0,-2.0,-1.0,0.0,-1.0,-1.0
+2021-07-02,0.0,-2.0,-1.0,0.0,-1.0,-1.0,-1.0
+2021-07-03,0.0,0.0,0.0,-1.0,-1.0,0.0,0.0
+2021-07-04,0.0,0.0,-1.0,-1.0,0.0,1.0,-2.0
+2021-07-05,0.0,0.0,-1.0,0.0,0.0,-2.0,-1.0
+2021-07-06,0.0,-1.0,0.0,0.0,-2.0,-1.0,-1.0
+2021-07-07,0.0,0.0,1.0,-1.0,0.0,-1.0,1.0
+2021-07-08,0.0,1.0,-1.0,0.0,-1.0,0.0,-2.0
+2021-07-09,0.0,-1.0,0.0,-1.0,-2.0,-1.0,1.0
+2021-07-10,0.0,0.0,0.0,0.0,-1.0,0.0,0.0
+2021-07-11,0.0,-1.0,-1.0,-1.0,0.0,0.0,-1.0
+2021-07-12,0.0,-1.0,-1.0,1.0,0.0,-4.0,-1.0
+2021-07-13,0.0,0.0,1.0,1.0,1.0,0.0,0.0
+2021-07-14,0.0,0.0,1.0,1.0,0.0,0.0,0.0
+2021-07-15,0.0,1.0,0.0,0.0,1.0,-1.0,-1.0
+2021-07-16,0.0,0.0,0.0,0.0,0.0,-1.0,-1.0
+2021-07-17,0.0,0.0,1.0,0.0,-1.0,-2.0,-3.0
+2021-07-18,0.0,1.0,0.0,0.0,-2.0,-2.0,1.0
+2021-07-19,0.0,-1.0,0.0,-2.0,-2.0,1.0,-2.0
+2021-07-20,0.0,0.0,-1.0,0.0,0.0,1.0,1.0
+2021-07-21,0.0,0.0,-1.0,1.0,0.0,1.0,0.0
+2021-07-22,0.0,-1.0,1.0,0.0,0.0,-1.0,0.0
+2021-07-23,0.0,1.0,0.0,0.0,0.0,1.0,3.0
+2021-07-24,0.0,-1.0,-1.0,0.0,1.0,2.0,1.0
+2021-07-25,0.0,0.0,-1.0,0.0,1.0,0.0,0.0
+2021-07-26,0.0,0.0,0.0,1.0,0.0,1.0,-2.0
+2021-07-27,0.0,0.0,1.0,0.0,2.0,-1.0,2.0
+2021-07-28,0.0,1.0,0.0,1.0,-2.0,1.0,
+2021-07-29,0.0,0.0,0.0,-1.0,1.0,,
+2021-07-30,0.0,0.0,0.0,0.0,,,
+2021-07-31,0.0,0.0,0.0,,,,
+2021-08-01,0.0,0.0,,,,,
+2021-08-02,0.0,,,,,,
