+++ static/data/forecast_dataframe.csv
@@ -227,3 +227,70 @@
 2021-05-25,20,15,15,14,14,15,16
 2021-05-26,15,15,15,13,14,16,16
 2021-05-27,15,14,14,14,15,16,16
+2021-05-28,14,14,15,15,16,17,16
+2021-05-29,14,14,16,16,17,17,16
+2021-05-30,14,16,16,17,16,15,15
+2021-05-31,16,17,18,17,15,16,16
+2021-06-01,16,18,18,14,15,16,16
+2021-06-02,18,18,15,16,16,17,14
+2021-06-03,17,14,15,16,17,13,13
+2021-06-04,14,15,16,18,15,13,13
+2021-06-05,14,17,18,13,13,14,15
+2021-06-06,17,18,13,13,16,16,16
+2021-06-07,18,13,13,15,16,16,16
+2021-06-08,13,14,15,16,15,15,16
+2021-06-09,14,16,15,16,16,15,16
+2021-06-10,16,15,15,16,15,15,15
+2021-06-11,16,16,16,15,16,15,14
+2021-06-12,16,16,16,16,16,15,14
+2021-06-13,15,16,16,15,15,14,14
+2021-06-14,15,16,14,15,13,14,14
+2021-06-15,16,15,15,14,14,14,15
+2021-06-16,14,15,15,15,14,15,15
+2021-06-17,15,14,15,14,15,15,15
+2021-06-18,14,15,14,15,15,15,15
+2021-06-19,15,15,14,15,15,16,16
+2021-06-20,15,15,16,16,15,16,14
+2021-06-21,15,16,15,16,14,14,14
+2021-06-22,16,15,16,14,14,14,14
+2021-06-23,15,16,14,14,14,15,17
+2021-06-24,16,14,13,14,15,16,17
+2021-06-25,13,13,15,15,16,17,16
+2021-06-26,13,15,15,16,17,17,15
+2021-06-27,15,15,16,17,16,13,13
+2021-06-28,15,16,17,16,13,11,13
+2021-06-29,16,16,16,13,12,13,14
+2021-06-30,16,16,16,12,12,14,14
+2021-07-01,17,15,12,14,14,14,14
+2021-07-02,16,12,14,14,14,14,15
+2021-07-03,10,13,14,14,14,14,15
+2021-07-04,13,14,14,14,14,14,15
+2021-07-05,14,13,14,14,15,15,16
+2021-07-06,13,14,14,15,15,16,15
+2021-07-07,13,14,14,14,15,15,14
+2021-07-08,14,14,14,15,15,15,16
+2021-07-09,15,14,15,15,17,15,16
+2021-07-10,13,15,14,15,15,17,15
+2021-07-11,15,15,16,15,17,15,14
+2021-07-12,14,16,15,16,15,17,15
+2021-07-13,15,14,16,14,12,14,15
+2021-07-14,14,17,14,12,14,15,13
+2021-07-15,17,14,13,14,14,14,14
+2021-07-16,15,13,14,15,13,14,13
+2021-07-17,13,14,14,13,14,14,14
+2021-07-18,14,14,13,13,14,13,13
+2021-07-19,15,14,13,14,13,13,14
+2021-07-20,13,13,13,11,14,11,15
+2021-07-21,13,12,12,13,12,15,17
+2021-07-22,12,12,13,12,16,18,16
+2021-07-23,11,13,12,16,17,15,14
+2021-07-24,14,13,17,17,15,15,14
+2021-07-25,12,16,18,16,16,15,19
+2021-07-26,16,17,16,16,15,18,17
+2021-07-27,17,16,16,15,17,16,14
+2021-07-28,16,16,15,18,17,15,13
+2021-07-29,17,15,19,16,15,14,15
+2021-07-30,15,19,15,16,15,15,15
+2021-07-31,19,15,16,13,14,15,16
+2021-08-01,15,16,13,14,16,16,17
+2021-08-02,16,14,14,15,16,16,17
