+++ static/data/last_row.csv
@@ -1,2 +1,2 @@
 ,extended_text
-0,"Cloudy. Very high (90%) chance of showers, most likely from late this morning. Winds westerly 20 to 30 km/h turning southwesterly 15 to 25 km/h in the late afternoon."
+0,Mostly sunny. Patchy fog in the northeast suburbs early this morning. Medium (50%) chance of showers in the northwest suburbs developing in the late evening. Winds northerly 25 to 40 km/h.
