+++ static/data/persistence_dataframe.csv
@@ -221,8 +221,75 @@
 2021-05-19,2,0.0,0.0,0.0,1.0,0.0,0.0
 2021-05-20,-1,0.0,0.0,0.0,-1.0,1.0,-1.0
 2021-05-21,-1,0.0,0.0,0.0,0.0,-1.0,0.0
-2021-05-22,-1,0.0,0.0,1.0,0.0,0.0,
-2021-05-23,1,0.0,0.0,0.0,0.0,,
-2021-05-24,-1,0.0,0.0,0.0,,,
-2021-05-25,5,0.0,0.0,,,,
-2021-05-26,0,0.0,,,,,
+2021-05-22,-1,0.0,0.0,1.0,0.0,0.0,0.0
+2021-05-23,1,0.0,0.0,0.0,0.0,0.0,0.0
+2021-05-24,-1,0.0,0.0,0.0,-1.0,-1.0,-1.0
+2021-05-25,5,0.0,0.0,0.0,0.0,-1.0,0.0
+2021-05-26,0,0.0,-1.0,1.0,0.0,0.0,0.0
+2021-05-27,1,0.0,0.0,0.0,1.0,0.0,2.0
+2021-05-28,0,0.0,-1.0,1.0,0.0,1.0,1.0
+2021-05-29,0,0.0,0.0,0.0,1.0,0.0,-2.0
+2021-05-30,-2,0.0,0.0,1.0,1.0,-1.0,-1.0
+2021-05-31,0,-1.0,0.0,0.0,-1.0,-2.0,1.0
+2021-06-01,-2,0.0,-1.0,0.0,-1.0,1.0,2.0
+2021-06-02,1,-1.0,-1.0,-2.0,1.0,1.0,-1.0
+2021-06-03,3,0.0,-1.0,1.0,1.0,0.0,1.0
+2021-06-04,0,-1.0,1.0,0.0,-2.0,1.0,3.0
+2021-06-05,-3,0.0,0.0,0.0,1.0,2.0,1.0
+2021-06-06,-1,0.0,0.0,1.0,0.0,0.0,0.0
+2021-06-07,5,0.0,1.0,1.0,0.0,0.0,-1.0
+2021-06-08,-1,0.0,1.0,0.0,1.0,0.0,-1.0
+2021-06-09,-2,0.0,1.0,0.0,-1.0,0.0,0.0
+2021-06-10,0,1.0,1.0,-1.0,0.0,1.0,-1.0
+2021-06-11,0,0.0,-1.0,0.0,0.0,-1.0,1.0
+2021-06-12,1,-1.0,-1.0,0.0,-2.0,0.0,0.0
+2021-06-13,0,-1.0,0.0,-1.0,0.0,0.0,1.0
+2021-06-14,-1,0.0,0.0,0.0,1.0,1.0,1.0
+2021-06-15,2,-1.0,0.0,0.0,1.0,1.0,0.0
+2021-06-16,-1,0.0,-1.0,0.0,1.0,0.0,1.0
+2021-06-17,1,0.0,0.0,1.0,0.0,1.0,0.0
+2021-06-18,-1,0.0,1.0,0.0,1.0,0.0,1.0
+2021-06-19,0,0.0,1.0,1.0,0.0,0.0,-3.0
+2021-06-20,0,0.0,0.0,-1.0,1.0,-3.0,-1.0
+2021-06-21,-1,0.0,0.0,0.0,-1.0,-1.0,1.0
+2021-06-22,1,0.0,0.0,-1.0,-1.0,1.0,1.0
+2021-06-23,-1,0.0,-1.0,-1.0,1.0,0.0,-1.0
+2021-06-24,3,-1.0,0.0,1.0,0.0,0.0,-1.0
+2021-06-25,0,0.0,0.0,0.0,0.0,-1.0,1.0
+2021-06-26,-2,0.0,0.0,0.0,-1.0,0.0,1.0
+2021-06-27,0,0.0,0.0,-1.0,1.0,3.0,-3.0
+2021-06-28,-1,0.0,-1.0,1.0,3.0,-1.0,0.0
+2021-06-29,0,0.0,1.0,3.0,-2.0,0.0,0.0
+2021-06-30,-1,1.0,0.0,-2.0,1.0,0.0,-1.0
+2021-07-01,1,1.0,-2.0,-1.0,0.0,-1.0,-1.0
+2021-07-02,6,-2.0,-1.0,0.0,-1.0,-1.0,-1.0
+2021-07-03,-3,0.0,0.0,-1.0,-1.0,0.0,0.0
+2021-07-04,-1,0.0,-1.0,-1.0,0.0,1.0,-2.0
+2021-07-05,1,0.0,-1.0,0.0,0.0,-2.0,-1.0
+2021-07-06,0,-1.0,0.0,0.0,-2.0,-1.0,-1.0
+2021-07-07,-1,0.0,1.0,-1.0,0.0,-1.0,1.0
+2021-07-08,-1,1.0,-1.0,0.0,-1.0,0.0,-2.0
+2021-07-09,2,-1.0,0.0,-1.0,-2.0,-1.0,1.0
+2021-07-10,-2,0.0,0.0,0.0,-1.0,0.0,0.0
+2021-07-11,1,-1.0,-1.0,-1.0,0.0,0.0,-1.0
+2021-07-12,-1,-1.0,-1.0,1.0,0.0,-4.0,-1.0
+2021-07-13,1,0.0,1.0,1.0,1.0,0.0,0.0
+2021-07-14,-3,0.0,1.0,1.0,0.0,0.0,0.0
+2021-07-15,2,1.0,0.0,0.0,1.0,-1.0,-1.0
+2021-07-16,2,0.0,0.0,0.0,0.0,-1.0,-1.0
+2021-07-17,-1,0.0,1.0,0.0,-1.0,-2.0,-3.0
+2021-07-18,-1,1.0,0.0,0.0,-2.0,-2.0,1.0
+2021-07-19,2,-1.0,0.0,-2.0,-2.0,1.0,-2.0
+2021-07-20,0,0.0,-1.0,0.0,0.0,1.0,1.0
+2021-07-21,1,0.0,-1.0,1.0,0.0,1.0,0.0
+2021-07-22,1,-1.0,1.0,0.0,0.0,-1.0,0.0
+2021-07-23,-3,1.0,0.0,0.0,0.0,1.0,3.0
+2021-07-24,2,-1.0,-1.0,0.0,1.0,2.0,1.0
+2021-07-25,-4,0.0,-1.0,0.0,1.0,0.0,0.0
+2021-07-26,-1,0.0,0.0,1.0,0.0,1.0,-2.0
+2021-07-27,1,0.0,1.0,0.0,2.0,-1.0,2.0
+2021-07-28,-1,1.0,0.0,1.0,-2.0,1.0,
+2021-07-29,2,0.0,0.0,-1.0,1.0,,
+2021-07-30,-4,0.0,0.0,0.0,,,
+2021-07-31,4,0.0,0.0,,,,
+2021-08-01,-1,0.0,,,,,
