+++ training_data/task1_training_weather.py
